feat: created OderService

+++ src/main/java/com/macariomiguel/ecommerceproject/dto/ProductRequestDTO.java
@@ -2,5 +2,7 @@ package com.macariomiguel.ecommerceproject.dto;
 
 import lombok.NonNull;
 
-public record ProductRequestDTO(String name, String description, Double price, Long category_id, @NonNull String sku) {
+import java.math.BigDecimal;
+
+public record ProductRequestDTO(String name, String description, BigDecimal price, Long category_id, @NonNull String sku) {
 }
